% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{calculateCovariance}
\alias{calculateCovariance}
\title{Calculating the covariance of a sequence of estimated treatment effects}
\usage{
calculateCovariance(
  data,
  estimationMethod,
  estimand = "difference",
  bootstraps = 2500,
  previousDatasets = list(),
  parametersPreviousEstimators = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing the observed data at a given time. This data frame should have the same structure as the data frame outputted by \code{data_at_time_t}.}

\item{estimationMethod}{A function naming the function to be called to estimate the treatment effect. Included in this package: \code{standardization} and \code{rctmle}.}

\item{estimand}{A character string indicating the estimand of interest; treatment effect on "difference" scale, "ratio" scale or odds ratio ("oddsratio") scale.}

\item{bootstraps}{The number of bootstraps to calculate the covariance.}

\item{previousDatasets}{A list of data frames with same structure as in output of \code{data_at_time_t}. They represent the observed datasets up to the previous analysis time.}

\item{parametersPreviousEstimators}{A list of a list with the parameters used for the previous estimates. One can also change the estimation method by adding \code{estimationMethod} to the different lists corresponding with the different analyses.}

\item{...}{Further arguments for the estimator function.}
}
\value{
A vector of the estimated covariance(s) between the current estimate and previous estimate(s) as well as the estimated variance of the current estimate.
}
\description{
A function calculating the covariance of the current and previous estimates of the tretment effect.
}
