% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{calculateEstimate}
\alias{calculateEstimate}
\title{Calculating an estimate based on a given dataset}
\usage{
calculateEstimate(data, estimationMethod, estimand = "difference", ...)
}
\arguments{
\item{data}{A data frame containing the observed data at a given time. This data frame should have the same structure as the data frame outputted by \code{data_at_time_t}.}

\item{estimationMethod}{A function naming the function to be called to estimate the treatment effect. Included in this package: \code{standardization} and \code{tmle}.}

\item{estimand}{A character string indicating the estimand of interest; treatment effect on "difference" scale, "ratio" scale or odds ratio ("oddsratio") scale.}

\item{...}{Further arguments for the estimator function.}
}
\value{
An estimated value for the treatment effect based on the observed dataset and estimation method.
}
\description{
A function estimating the treatment effect for a given data frame and estimator.
}
