% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{calculateVariance}
\alias{calculateVariance}
\title{Calculating the variance of the estimated treatment effect}
\usage{
calculateVariance(
  data,
  estimationMethod,
  estimand = "difference",
  bootstraps = 2500,
  ...
)
}
\arguments{
\item{data}{A data frame containing the observed data at a given time. This data frame should have the same structure as the data frame outputted by \code{data_at_time_t}.}

\item{estimationMethod}{A function naming the function to be called to estimate the treatment effect. Included in this package: \code{standardization} and \code{tmle}.}

\item{estimand}{A character string indicating the estimand of interest; treatment effect on "difference" scale, "ratio" scale or odds ratio ("oddsratio") scale.}

\item{bootstraps}{The number of bootstraps to calculate the variance.}

\item{...}{Further arguments for the estimator function.}
}
\value{
An estimated value for the variance of the estimated treatment effect based on the observed dataset and estimation method.
}
\description{
A function that calculates the variance of the estimated treatment effect.
}
