% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_at_time_t.R
\name{data_at_time_t}
\alias{data_at_time_t}
\title{Determining dataset at a given analysis time}
\usage{
data_at_time_t(
  data,
  id_column,
  enrollment_time,
  treatment_column,
  covariate_columns,
  outcome_columns,
  outcome_times,
  analysis_time
)
}
\arguments{
\item{data}{A data frame.}

\item{id_column}{A character string representing the column name of participant identifiers.}

\item{enrollment_time}{A character string representing the column name of the enrollment times.}

\item{treatment_column}{A character string representing the column name of treatment variable.}

\item{covariate_columns}{A vector of character string representing the column names of the covariates.}

\item{outcome_columns}{A vector of character string representing the column names of the outcomes (including intermediate outcomes).}

\item{outcome_times}{A vector of character strings representing the column names of the outcome times (including intermediate outcomes).}

\item{analysis_time}{A numeric value for the time of the analysis.}
}
\value{
A data frame with the data of the patients already in the study at \code{analysis_time}. In addition to the original dataset, the output dataset has the additional column(s):
\describe{
\item{\code{.r_j}}{which equals \code{1} if outcome j is already observed for participant j at \code{analysis_time} and \code{NA} otherwise.}
}
}
\description{
A function that returns the a data frame with the observed data at a given \code{analysis_time}.
}
