% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drordinal.R
\name{drordinal}
\alias{drordinal}
\title{Estimating estimands for ordinal endpoints}
\usage{
drordinal(
  data,
  estimand = c("weighted_mean", "log_odds", "mann_whitney")[1],
  outcome_column,
  treatment_column,
  covariate_columns,
  ...
)
}
\arguments{
\item{data}{A data frame containing the observed data at a given time.}

\item{estimand}{A character string indicating the estimand of interest (for an ordinal endpoint); difference in (weighted) means ("weighted_mean"), log odds ratio ("log_odds") and Mann-Whitney ("mann_whitney").}

\item{...}{Further arguments for the estimator function drord.}

\item{out}{See \code{drord} in package \code{drord}.}

\item{treat}{See \code{drord} in package \code{drord}.}

\item{covar}{See \code{drord} in package \code{drord}.}
}
\value{
An estimated value for the treatment effect based on the observed dataset and estimation method.
}
\description{
A function that estimates the treatment effect wrt an ordinal outcome via a targeted maximum likelihood estimator.
}
\examples{
ctn03_sim$arm_num = 1*(ctn03_sim$arm == "7-day")
ctn03_sim$cows_category_eot_num = as.numeric(ctn03_sim$cows_category_eot)
drordinal(data = ctn03_sim,
estimand = c("weighted_mean"),
outcome_column = "cows_category_eot_num",
treatment_column = "arm_num",
covariate_columns = c("stability_dose",
                     "arsw_score_bl", "cows_total_score_bl",
                     "vas_crave_opiates_bl",
                     "vas_current_withdrawal_bl",
                     "vas_study_tx_help_bl",
                     "uds_opioids_bl", "uds_oxycodone_bl",
                     "uds_any_positive_bl"))

}
