% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{interimDecision}
\alias{interimDecision}
\title{Making a decision based on a given test statistic.}
\usage{
interimDecision(
  testStatistic,
  alpha = 0.05,
  beta = 0.2,
  alternative = "two.sided",
  typeOfDesign = "asOF",
  typeBetaSpending = "none",
  plannedAnalyses,
  plannedInformationTimes = NULL,
  previousInformationTimes,
  currentInformationTime,
  futilityStopping = FALSE,
  ...
)
}
\arguments{
\item{testStatistic}{A numeric value for the test statistic corresponding with the current analysis.}

\item{alpha}{The (total/cumulative) siginificance level alpha.}

\item{beta}{Type II error rate.}

\item{alternative}{A character string specifying the alternative hypothesis. The alternative is one of the following: two-sided (\code{two.sided}; default), one-sided less (\code{less}) or one-sided more (\code{more}).}

\item{typeOfDesign}{The type of design (see \code{getDesignGroupSequential} in \code{rpact} package).}

\item{typeBetaSpending}{The type of beta spending (see \code{getDesignGroupSequential} in \code{rpact} package).}

\item{plannedAnalyses}{A numeric value specifying the number of planned analyses in the group sequential design.}

\item{plannedInformationTimes}{A vector of numeric values representing the information times at which analyses (interim and final) are planned. This only needs to be specified if \code{futilityStopping=TRUE}.}

\item{previousInformationTimes}{A vector or numeric values representing the information times of the estimates up to the previous analysis.}

\item{currentInformationTime}{The information time corresponding with the estimate at the current analysis.}

\item{futilityStopping}{\code{futilityStopping=TRUE} if futility stopping is allowed. Default is \code{FALSE}. Note that futility stopping is only possible for one-sided testing.}

\item{...}{Further arguments for the arguments of \code{getDesignGroupSequential} in \code{rpact} package.}
}
\value{
A string value indicating whether one can stop for \code{"Futility"}, \code{"Efficacy"} or whether one can not stop (\code{"No"}).
}
\description{
A function that makes a decision based on a given test statistic corresponding with the current analysis.
}
