% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interimInformation.R
\name{interimInformation}
\alias{interimInformation}
\title{Estimating the (statistical) information for a given dataset and estimation method}
\usage{
interimInformation(
  data,
  totalInformation,
  analysisNumber,
  previousDatasets = list(),
  estimationMethod,
  estimand = "difference",
  update,
  previousEstimatesOriginal = c(),
  previousCovMatrixOriginal = c(),
  parametersPreviousEstimators = NULL,
  correction = "no",
  ...
)
}
\arguments{
\item{data}{A data frame containing the observed data at a given time. This data frame should have the same structure as the data frame outputted by \code{data_at_time_t}.}

\item{totalInformation}{A numeric value indicating the total/maximum information.}

\item{analysisNumber}{A numeric value representing the number of the current analysis.}

\item{previousDatasets}{A list of data frames with same structure as in output of \code{data_at_time_t}. They represent the observed datasets up to the previous analysis time.}

\item{estimationMethod}{A function naming the function to be called to estimate the treatment effect. Included in this package: \code{standardization} and \code{tmle}.}

\item{estimand}{A character string indicating the estimand of interest; treatment effect on "difference" scale, "ratio" scale or odds ratio ("oddsratio") scale.}

\item{update}{A character string indicating whether information needs to be calculated based on the updated estimate ("yes") or based on the original estimate ("no").}

\item{previousEstimatesOriginal}{A vector of numeric values containing the (original) estimates up to the previous analysis.}

\item{previousCovMatrixOriginal}{A matrix of numeric varlues respresenting the covariance matrix of the (original) estimates up to the previous analysis.}

\item{parametersPreviousEstimators}{A list of a list with the parameters used for the previous estimates. One can also change the estimation method by adding \code{estimationMethod} to the different lists corresponding with the different analyses.}

\item{correction}{Should a small sample correction be included?}

\item{...}{Further arguments for the estimator function or calculation of the variance (ie, number of bootstraps).}
}
\value{
An object of the class \code{interimInformation}.
\describe{
\item{information}{The (estimated) information available in the data frame \code{data} for the treatment effect of interest.}
\item{informationTime}{The information time, that is, the (estimated) information available relative to the total information.}}
}
\description{
A function estimating the (statistical) information based on a data frame and estimation function.
}
