% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{standardization}
\alias{standardization}
\title{Standardization estimator}
\usage{
standardization(
  data,
  estimand = "difference",
  y0_formula,
  y1_formula,
  family,
  treatment_column = NULL
)
}
\arguments{
\item{data}{A data frame containing the observed data at a given time. This data frame should have the same structure as the data frame outputted by \code{data_at_time_t}.}

\item{estimand}{A character string indicating the estimand of interest; treatment effect on "difference" scale, "ratio" scale or odds ratio ("oddsratio") scale.}

\item{y0_formula}{An object of the class \code{formula}, describing the model to be fitted for the outcome under control.}

\item{y1_formula}{An object of the class \code{formula}, describing the model to be fitted for the outcome under treatment.}

\item{family}{A description of the error distribution and link function to be used in the model (see \code{family}).}

\item{treatment_column}{A character string representing the column name of treatment variable.}
}
\value{
An object of the class \code{standardization}.
\describe{
\item{estimate}{Original estimate based on a standardization estimator (also called G-computation).}
\item{y1_pred}{Predictions of the outcome under treatment for all recruited participants.}
\item{y0_pred}{Predictions of the outcome under control for all recruited participants.}}
}
\description{
A function that estimates the treatment effect via a standardization estimator (also called G-computation).
}
