% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survrct_diff.R
\name{survrct_diff}
\alias{survrct_diff}
\title{Estimating time-to-event estimands}
\usage{
survrct_diff(
  data,
  estimand = c("rmst", "survprob")[1],
  outcome.formula,
  trt.formula,
  coarsen = 1,
  algo = c("glm", "lasso", "rf", "xgboost"),
  crossfit = TRUE,
  horizon
)
}
\arguments{
\item{data}{A data frame containing the observed data at a given time.}

\item{estimand}{A character string indicating the estimand of interest (for a time-to-event endpoint); difference in restricted mean survival time ("rmst") or difference in survival probabilities ("survprob").}

\item{outcome.formula}{See \code{survrct} in package \code{adjrct}.}

\item{trt.formula}{See \code{survrct} in package \code{adjrct}.}

\item{coarsen}{See \code{survrct} in package \code{adjrct}.}

\item{algo}{See \code{survrct} in package \code{adjrct}.}

\item{crossfit}{See \code{survrct} in package \code{adjrct}.}

\item{horizon}{See \code{rmst} and \code{survprob} in package \code{adjrct}.}
}
\value{
An estimated value for the treatment effect based on the observed dataset and estimation method.
}
\description{
A function that estimates the treatment effect wrt a time-to-event outcome via a targeted maximum likelihood estimator.
}
\examples{
colon_cancer$trt = ifelse(colon_cancer$arm=="Obs", 1, 0)
survrct_diff(data=colon_cancer,
estimand = "survprob",
outcome.formula=Surv(months_to_death, event_death) ~ trt + age,
trt.formula=trt ~ 1,
horizon=14)

}
