% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{tmle}
\alias{tmle}
\title{Targeted maximum likelihood estimator}
\usage{
tmle(
  data,
  estimand = "difference",
  propensity_score_formula,
  inverse_weight_formulas,
  outcome_formulas,
  outcome_type = c("gaussian", "logistic", "binomial", "multinomial-binomial")[1],
  outcome_range = NULL,
  absorbing_state = NULL,
  absorbing_outcome = NULL,
  impute_formulas = NULL,
  impute_model = c("gaussian", "binomial", "beta", "pmm", "multinomial")[1],
  imputation_args = NULL,
  ci = FALSE,
  verbose = FALSE,
  bootstrap_n = 10000,
  bootstrap_type = c("bca", "norm", "basic", "perc")[1],
  alpha = 0.05,
  ...
)
}
\arguments{
\item{data}{A data frame containing all observed data for the current analysis}

\item{estimand}{Estimand of interest; treatment effect on "difference" scale, "ratio" scale or odds ratio ("oddsratio") scale}

\item{propensity_score_formula}{TODO}

\item{inverse_weight_formulas}{TODO}

\item{outcome_formulas}{TODO}

\item{outcome_type}{TODO}

\item{outcome_range}{TODO}

\item{absorbing_state}{TODO}

\item{absorbing_outcome}{TODO}

\item{impute_formulas}{TODO}

\item{impute_model}{TODO}

\item{imputation_args}{TODO}

\item{ci}{TODO}

\item{verbose}{TODO}

\item{bootstrap_n}{TODO}

\item{bootstrap_type}{TODO}

\item{alpha}{TODO}

\item{...}{Further arguments for the estimator function.}
}
\value{
An object of class \code{standardization}.
\describe{
\item{estimate}{Original estimate based on a standardization estimator (also called G-computation).}
\item{y1_pred}{Predictions of the outcome under treatment for all recruited participants.}
\item{y0_pred}{Predictions of the outcome under control for all recruited participants.}}
}
\description{
A function that estimates the treatment effect via a (longitudinal) targeted maximum likelihood estimator.
}
