% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{updateEstimate}
\alias{updateEstimate}
\title{Orthogonalizing the original sequence of estimates up to the current analysis}
\usage{
updateEstimate(covMatrixOriginal, estimatesOriginal)
}
\arguments{
\item{covMatrixOriginal}{A matrix of numeric varlues respresenting the covariance matrix of the (original) estimates up to the previous analysis.}

\item{estimatesOriginal}{A vector of numeric values containing the (original) estimates up to the previous analysis.}
}
\value{
An object of the class \code{updateEstimate}.
\describe{
\item{estimateUpdated}{The updated/orthogonalized estimate at the current analysis.}
\item{varianceUpdated}{The estimated variance of the updated/orthogonalized estimate at the current analysis.}}
}
\description{
A function orthogonalizing the original sequence of estimates up to the current analysis.
}
