% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interimAnalysis.R
\name{interimAnalysis}
\alias{interimAnalysis}
\title{Performing an (interim) analysis based on a given dataset}
\usage{
interimAnalysis(
  data,
  totalInformation,
  estimationMethod,
  estimand = "difference",
  previousEstimatesOriginal = c(),
  previousCovMatrixOriginal = c(),
  previousInformationTimesOriginal = c(),
  previousInformationTimesUpdated = c(),
  previousDatasets = list(),
  null.value = 0,
  alpha = 0.05,
  beta = 0.2,
  alternative = "two.sided",
  typeOfDesign = "asOF",
  typeBetaSpending = "none",
  futilityStopping = FALSE,
  plannedAnalyses,
  plannedInformationTimes = NULL,
  parametersPreviousEstimators = NULL,
  correction = "no",
  ...
)
}
\arguments{
\item{data}{A data frame containing the observed data at a given time. This data frame should have the same structure as the data frame outputted by \code{data_at_time_t}.}

\item{totalInformation}{A numeric value indicating the total/maximum information.}

\item{estimationMethod}{A function naming the function to be called to estimate the treatment effect. Included in this package: \code{standardization} and \code{tmle}.}

\item{estimand}{A character string indicating the estimand of interest; treatment effect on "difference" scale, "ratio" scale or odds ratio ("oddsratio") scale.}

\item{previousEstimatesOriginal}{A vector of numeric values containing the (original) estimates up to the previous analysis.}

\item{previousCovMatrixOriginal}{A matrix of numeric varlues respresenting the covariance matrix of the (original) estimates up to the previous analysis.}

\item{previousInformationTimesOriginal}{A vector or numeric values representing the information times of the original estimates up to the previous analysis.}

\item{previousInformationTimesUpdated}{A vector or numeric values representing the information times of the updated estimates up to the previous analysis.}

\item{previousDatasets}{A list of data frames with same structure as in output of \code{data_at_time_t}. They represent the observed datasets up to the previous analysis time.}

\item{null.value}{A numeric value indicating the value of the treatment effect under the null hypothesis.}

\item{alpha}{The (total) siginificance level alpha.}

\item{beta}{Type II error rate.}

\item{alternative}{A character string specifying the alternative hypothesis. The alternative is one of the following: two-sided (\code{two.sided}; default), one-sided less (\code{less}) or one-sided more (\code{more}).}

\item{typeOfDesign}{The type of design (see \code{getDesignGroupSequential} in \code{rpact} package).}

\item{typeBetaSpending}{The type of beta spending (see \code{getDesignGroupSequential} in \code{rpact} package).}

\item{futilityStopping}{\code{futilityStopping=TRUE} if futility stopping is allowed. Default is \code{FALSE}. Note that futility stopping is only possible for one-sided testing.}

\item{plannedAnalyses}{A numeric value specifying the number of planned analyses in the group sequential design.}

\item{plannedInformationTimes}{A vector of numeric values representing the information times at which analyses (interim and final) are planned. This only needs to be specified if \code{futilityStopping=TRUE}.}

\item{parametersPreviousEstimators}{A list of a list with the parameters used for the previous estimates. One can also change the estimation method by adding \code{estimationMethod} to the different lists corresponding with the different analyses.}

\item{correction}{Should a small sample correction be included?}

\item{...}{Further arguments for the estimator function,calculation of variance/covariance (ie, number of bootstraps) or arguments for \code{getDesignGroupSequential} in \code{rpact} package.}
}
\value{
An object of the class \code{interimAnalysis}.
\describe{
\item{estimateOriginal}{The original estimate at the current analysis.}
\item{standardErrorOriginal}{The estimated standard error of original estimate at the current analysis.}
\item{testStatisticOriginal}{The test statistic corresponding with the original estimate at the current analysis.}
\item{decisionOriginal}{The decision corresponding with the original estimate at the current analysis.}
\item{informationTimeOriginal}{The information time corresponding with the original estimate at the current analysis.}
\item{estimateUpdated}{The updated/orthogonalized estimate at the current analysis.}
\item{standardErrorUpdated}{The estimated standard error of the updated/orthogonalized estimate at the current analysis.}
\item{testStatisticUpdated}{The test statistic corresponding with the updated estimate at the current analysis.}
\item{decisionUpdated}{The decision corresponding with the updated estimate at the current analysis: A string value indicating whether one can stop for \code{"Futility"}, \code{"Efficacy"} or whether one can not stop (\code{"No"}).}
\item{informationTimeUpdated}{The information time corresponding with the updated estimate at the current analysis.}
\item{covMatrixOriginal}{A matrix of numeric values representing the covariance matrix of the original estimates up to the current analysis.}}
}
\description{
A function performing an (interim) analysis based on a given data frame for a chosen estimator.
}
\examples{
colon_cancer_enr = colon_cancer
colon_cancer_enr$trt = ifelse(colon_cancer_enr$arm=="Obs", 1, 0)
colon_cancer_enr$enrollmentTime = c(rep(1:100, 9), 1:29)
colon_cancer_enr$deathTime = colon_cancer_enr$months_to_death +
colon_cancer_enr$enrollmentTime
analysis_dataset = data_at_time_t(
data = colon_cancer_enr,
id_column = "id",
analysis_time = 50,
enrollment_time = "enrollmentTime",
treatment_column = "trt",
covariate_columns = c("age" , "sex"),
outcome_columns = c("event_death"),
outcome_times =  c("deathTime"),
time_to_event = TRUE
)
interim1=interimAnalysis(data = analysis_dataset,
totalInformation=5000,
estimationMethod= survrct_diff,
estimand = "survprob",
previousEstimatesOriginal=c(),
previousCovMatrixOriginal=c(),
previousInformationTimesOriginal=c(),
previousInformationTimesUpdated=c(),
previousDatasets = list(),
null.value = 0,
alpha = 0.05,
beta = 0.2,
alternative = "two.sided",
typeOfDesign = "asP",
typeBetaSpending = "none",
futilityStopping = FALSE,
plannedAnalyses=2,
plannedInformationTimes = NULL,
parametersPreviousEstimators = NULL,
correction="no",
bootstraps =100,
outcome.formula=Surv(.time_to_event_1, event_death) ~ trt + age,
trt.formula=trt ~ 1,
horizon = 14)
interim1
final = interimAnalysis(data = colon_cancer,
                        totalInformation=5000,
                        estimationMethod= survrct_diff,
                        estimand = "survprob",
                        previousEstimatesOriginal=as.numeric(interim1[[1]]),
                        previousCovMatrixOriginal=interim1[[11]],
                        previousInformationTimesOriginal=as.numeric(interim1[[5]]),
                        previousInformationTimesUpdated=as.numeric(interim1[[10]]),
                        previousDatasets = list(analysis_dataset),
                        null.value = 0,
                        alpha = 0.05,
                        beta = 0.2,
                        alternative = "two.sided",
                        typeOfDesign = "asP",
                        typeBetaSpending = "none",
                        futilityStopping = FALSE,
                        plannedAnalyses=2,
                        bootstraps=10,
                        plannedInformationTimes = NULL,
                        parametersPreviousEstimators = list(list(outcome.formula=Surv(.time_to_event_1, event_death) ~ trt + age,
                                                             trt.formula=trt ~ 1,
                                                             horizon = 14)),
                        correction="no",
                        outcome.formula=Surv(months_to_death, event_death) ~ trt + age,
                        trt.formula=trt ~ 1,
                        horizon = 14)
final

}
