% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{tmle}
\alias{tmle}
\title{Targeted maximum likelihood estimator}
\usage{
tmle(
  data,
  estimand = "difference",
  propensity_score_formula,
  inverse_weight_formulas,
  outcome_formulas,
  outcome_type = c("gaussian", "logistic", "binomial", "multinomial-binomial")[1],
  outcome_range = NULL,
  absorbing_state = NULL,
  absorbing_outcome = NULL,
  impute_formulas = NULL,
  impute_model = c("gaussian", "binomial", "beta", "pmm", "multinomial")[1],
  imputation_args = NULL,
  ci = FALSE,
  verbose = FALSE,
  bootstrap_n = 10000,
  bootstrap_type = c("bca", "norm", "basic", "perc")[1],
  alpha = 0.05,
  ...
)
}
\arguments{
\item{data}{A data frame containing the observed data at a given time. This data frame should have the same structure as the data frame outputted by \code{data_at_time_t}.}

\item{estimand}{A character string indicating the estimand of interest; treatment effect on "difference" scale, "ratio" scale or odds ratio ("oddsratio") scale.}

\item{propensity_score_formula}{A formula for the propensity score model - the left-hand-side indicates the binary treatment assignment indicator variable.}

\item{inverse_weight_formulas}{A list of formulas for computing inverse probability of censoring weights. Only the right-hand side of the formula is used.}

\item{outcome_formulas}{A list of formulas with one formula per outcome model, ordered from the first to the last observed outcome.}

\item{outcome_type}{A character string indicating the type of outcome model to be fitted: "gaussian", "logistic", "binomial", "multinomial-binomial".}

\item{outcome_range}{A numeric vector of length 2 indicating the range of the outcome. Only used when outcome_type == "logistic".}

\item{absorbing_state}{A character vector indicating which level (if any) of the outcome is an absorbing state. Only used when outcome_type == "multinomial-binomial".}

\item{absorbing_outcome}{A value of 0 or 1 denoting whether the absorbing state is a success (1) or failure (0) for the binary final outcome. Only used when outcome_type == "multinomial-binomial".}

\item{impute_formulas}{A list of formulas with one formula per intermediate outcome, containing the paramaterization of imputation models for intermediate outcomes.}

\item{impute_model}{A character vector indicating the family of imputation model to use: "gaussian" (default), "binomial", "beta", "pmm", "multinomial".}

\item{imputation_args}{A list of arguments to pass to functions performing imputation.}

\item{ci}{TRUE/FALSE - Should bootstrap-based confidence intervals be returned (TRUE) or just the estimate (FALSE).}

\item{verbose}{TRUE/FALSE - Should all models and intermediate results be returned (TRUE) or just the estimate (FALSE).}

\item{bootstrap_n}{Number of bootstrap replicates used to perform bootstrap inference.}

\item{bootstrap_type}{The type of bootstrap to perform: "bca" (default), "norm", "basic", or "perc".}

\item{alpha}{Type I Error rate used in the confidence level for bootstrap inference - Defaults to 0.05.}

\item{...}{Further arguments for the estimator function.}
}
\value{
An object of class \code{standardization}.
\describe{
\item{estimate}{Original estimate based on a standardization estimator (also called G-computation).}
\item{y1_pred}{Predictions of the outcome under treatment for all recruited participants.}
\item{y0_pred}{Predictions of the outcome under control for all recruited participants.}}
}
\description{
A function that estimates the treatment effect via a (longitudinal) targeted maximum likelihood estimator.
}
